requirements.txt
certifi==2019.3.9
googleapis-common-protos==1.5.9
grpcio==1.19.0
grpcio-tools==1.16.1
numpy==1.16.2
pandas==0.24.0
protobuf==3.7.1
pymortar==1.0.4
python-dateutil==2.8.0
pytz==2017.2
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
xbos-services-getter==0.0.2.dev17
